fix(federated): enforce declared-type fit at write, project EAV narrow ints by storage width (#384) (#500)
* fix(transform): reject EAV numeric-family values that do not fit the declared integer type (#384)

smallint/integer/bigint EAV values are stored in unconstrained NUMERIC, so
the write funnel is the only place declared width can be enforced; values
that do not fit (out of range or non-integral) now answer 400 instead of
being silently stored and then diverging across query tiers.

* fix(sqlgen,cdc,federated): project EAV-only integer/smallint by storage width (BIGINT) on every DuckDB tier (#384)

eav_data.value_numeric is unconstrained NUMERIC, so casting the DuckDB
projection by declared width TRY_CAST a stored out-of-range value to NULL
on every DuckDB leg while hot Postgres matched it. All four projection
surfaces move in lockstep: the hot EAV pivot (buildEAVPivotExpr), list
element casts (eavElementCastExpr), the CDC parquet export (castEAVValue),
and the cold NULL-augmentation type (DuckDBNullScanType, now
binding-aware). Column-bound integer/smallint keep declared width: their
physical storage is int4/int2. Pre-change parquet files carrying INT32
columns promote losslessly to BIGINT via the existing union_by_name scans
(read + compaction verified).

* fix(sqlgen): cast DuckDB predicate operands by storage width — BIGINT for narrow ints, DOUBLE for numeric (#384)

The strict CAST(? AS INTEGER/SMALLINT) raised a DuckDB Conversion Error
for operands past the declared range while Postgres answered the same
filter; CAST(? AS DECIMAL(38,10)) silently truncated numeric operands
beyond 10 fractional digits and overflowed past ~1e28 against columns
that are DOUBLE on every tier. Operands now cast at the width the columns
actually carry, so out-of-range operands mismatch identically on both
engines.

* fix(sqlgen): PG EAV bool predicates compare value_numeric <> 0 truthiness, matching every DuckDB tier (#384)

The OLTP route bound bool equality as value_numeric = 1.0/0.0 while every
DuckDB leg derives the column as value_numeric <> 0, so a stored 2 matched
equals:1 on the federated tiers and not on the OLTP route. Both EAV EXISTS
emitters (pgEavTypedEmitter and the hybrid builder) now render the same
truthiness expression via PgEavLeafPayload.ComparisonLHS with a BOOLEAN
bind.

* test(e2e): pin hot/cold agreement for out-of-declared-width EAV values (#384)

Four production-harness acceptance tests: integer/smallint values past the
declared width (planted via direct eav_data updates, standing in for
pre-rejection history) answer identically on the hot Postgres route and
the cold DuckDB tier; the write path rejects new out-of-width values as
invalid input; bool equality follows the value_numeric <> 0 truth rule on
both routes; numeric operands with >10 fractional digits or magnitudes
past 1e28 compare instead of truncating or erroring.

* docs(federated-query): record the numeric width contract (#384)

* style(sqlgen): gofmt predicate characterization matrix

* test(e2e): widen EAV-only integer parquet shape pins to storage width (#384)

The full-type, schema-evolution, compaction-evolution, and manifest-stamp
suites pinned INTEGER/SMALLINT as the physical parquet type of EAV-only
integer attributes; the export now writes those columns at storage width
BIGINT.

* docs(sqlgen): record the lists-are-EAV-only premise behind LIST element operand widths (#384)

Review O1 on PR #500: the list operand type rides MapValueTypeToDuckDBType
and therefore EAV storage width, which holds because list attributes cannot
bind a main column; a future list column binding needs an explicit
storage-vs-declared width decision here. Also fixes the stale
LIST(INTEGER) example.

* fix(sqlgen,cdc,federated): EAV numeric storage width is DOUBLE; round-trip float64 operands; one bool operand rule (#384)

Resolves the four findings of the second PR #500 review:

P1a — the BIGINT projection for EAV-only integer/smallint mangled
non-integral history (TRY_CAST(1.5 AS BIGINT) = 2), matching equals:2 only
on the DuckDB legs while Postgres compared the raw NUMERIC 1.5. The true
storage width is DOUBLE: the write funnel narrows every numeric-family
value through float64 (numutil.Float64), so value_numeric only ever holds
float64 images and DOUBLE reproduces all of them — over-range and
non-integral alike — exactly. All four projection surfaces (hot pivot,
list elements, CDC export, cold NULL augmentation) move from BIGINT to
DOUBLE in lockstep, joining the numeric class.

P2a — integer/smallint operands follow to CAST(? AS DOUBLE), so an
operand of any magnitude (1e30) compares on both engines instead of
raising a DuckDB Conversion Error. The bigint sibling (BIGINT must stay
exact above 2^53) is pre-existing and tracked as #502.

P1b — decimalString now renders float64 operands in shortest round-trip
form (strconv.FormatFloat 'g' -1); %.15g dropped the 16th-17th
significant digits, so DuckDB bound an adjacent float64 and equality
missed rows only on that route.

P2b — bool operands parse under one engine-shared rule (parseBoolOperand:
ParseBool spellings, else any integer at the PG path's historical >0
truthiness). Before, "2" bound only on PG and "true" only on DuckDB.

The acceptance e2e now drives every probe through hot, warm (real CDC
flush), and cold (real compaction) and adds non-integral-history,
16-digit numeric, bool-spelling, and past-BIGINT-operand probes. Suite
shape pins and docs §6.4 updated to the DOUBLE contract.

* test(e2e): scan EAV-only level/qty as DOUBLE in the wide parquet scanner (#384)

The physical-layer scanner read the parquet columns into
NullInt16/NullInt32; with the columns at storage width DOUBLE, large
values render in scientific notation through database/sql's string
conversion and the scan errored. Scan as NullFloat64 and compare against
the declared-width truth (in-range integers are exact in DOUBLE).

* fix(sqlgen): narrow integral EAV operands through the write funnel's float64 for DOUBLE-width classes (#384)

Fourth-review P1: with the DOUBLE projection, an integral operand above
2^53 still bound as exact int64 on the Postgres EAV routes while DuckDB's
CAST(? AS DOUBLE) rounded it — so equals:9007199254740993 missed a stored
float64 image 9007199254740992 on hot Postgres and matched it on every
DuckDB tier. NarrowEAVNumericOperand converts integral operands for
integer/smallint/numeric EAV comparisons to float64 — the same narrowing
the stored data took — in both the predicate normalizer and the batch
attr-value anchor (which must reach the same verdict on the same literal,
#355). bigint keeps exact int64 binds: its storage is BIGINT on every
tier. Docs §6.4 wording tightened; 2^53-boundary probes added to the
acceptance e2e on all three tiers.

* docs(sqlgen): note the Duck-side exact-int64/CAST-rounding parity at the integer passthrough (#384)

Fifth-review O1: the DuckDB binder keeps the exact int64 and relies on
CAST(? AS DOUBLE) to round in-engine, landing on the same float64 image
the PG binders reach via NarrowEAVNumericOperand — record that at the
passthrough site so the parity reasoning is discoverable where the two
sides appear to diverge.

* docs(federated-query): make §6.3's narrow-int casting line defer to the §6.4 binding-dependent contract (#384)

Sixth-review observation: the generic Smart Type Casting bullet still read
'smallint → INTEGER or BIGINT', which contradicts §6.4's precise ruling
(column-bound keeps physical width, EAV-only carries storage width
DOUBLE).

+++ internal/e2e_harness/production/compaction_evolution_e2e_test.go
@@ -66,7 +66,7 @@ func buildEvoV1Profile() AttrProfile {
 }
 
 // buildEvoV2Profile seeds v2 rows: new_col + fractional score, so any silent
-// DOUBLE→INTEGER coercion in the merge corrupts a visible value.
+// DOUBLE→integer coercion in the merge corrupts a visible value.
 func buildEvoV2Profile() AttrProfile {
 	return buildEvolutionProfile(func(ordinal int) map[string]any {
 		return map[string]any{
@@ -78,7 +78,7 @@ func buildEvoV2Profile() AttrProfile {
 
 // buildEvolutionEquivalenceQueries is the before/after snapshot set: an
 // unsorted page, a sort on a generation-stable attribute, a score filter
-// spanning both generations (v1 INTEGER rows and v2 DOUBLE rows in one
+// spanning both generations (v1 integer-generation rows and v2 DOUBLE rows in one
 // numeric domain, with the boundary row score=20 included), and a new_col
 // filter that only v2-generation rows can match (v1 rows are NULL).
 func buildEvolutionEquivalenceQueries(schema SchemaRef) []Query {
@@ -179,8 +179,8 @@ func assertMergedBaseUnion(ctx context.Context, t *testing.T, env *Env, key stri
 		"name":    "VARCHAR",
 		"value":   "DOUBLE",
 		"old_col": "VARCHAR", // v1 legacy column survives the union
-		"new_col": "INTEGER", // v2 addition present
-		"score":   "DOUBLE",  // INTEGER widened to the delta's DOUBLE
+		"new_col": "DOUBLE",  // v2 addition present
+		"score":   "DOUBLE",  // both generations export DOUBLE since #384; was BIGINT widened to the delta's DOUBLE
 	})
 
 	path := buildParquetS3Path(env, key)
@@ -291,11 +291,11 @@ func TestCompactionMixedGenerationEquivalence(t *testing.T) {
 	// shapes the equivalence pass proves nothing about cross-generation merge.
 	baseCols := describeParquetCols(ctx, t, env, seed.baseKey)
 	requireParquetCols(t, "base (v1)", baseCols, map[string]string{
-		"name": "VARCHAR", "value": "DOUBLE", "old_col": "VARCHAR", "score": "INTEGER"})
+		"name": "VARCHAR", "value": "DOUBLE", "old_col": "VARCHAR", "score": "DOUBLE"})
 	forbidParquetCols(t, "base (v1)", baseCols, "new_col")
 	deltaCols := describeParquetCols(ctx, t, env, seed.deltaKey)
 	requireParquetCols(t, "delta (v2)", deltaCols, map[string]string{
-		"score": "DOUBLE", "new_col": "INTEGER"})
+		"score": "DOUBLE", "new_col": "DOUBLE"})
 	forbidParquetCols(t, "delta (v2)", deltaCols, "old_col")
 
 	// Positive control for the row-level LWW assertion: the update targets'
@@ -395,7 +395,7 @@ func verifyPostCompactionEvolution(ctx context.Context, t *testing.T, env *Env, 
 
 	// Monotonic healing: the second merge output keeps the union shape.
 	requireParquetCols(t, "second merged base", describeParquetCols(ctx, t, env, second.NewBaseKey),
-		map[string]string{"old_col": "VARCHAR", "new_col": "INTEGER", "score": "DOUBLE"})
+		map[string]string{"old_col": "VARCHAR", "new_col": "DOUBLE", "score": "DOUBLE"})
 	path := buildParquetS3Path(env, second.NewBaseKey)
 	var oldColSurvivors int
 	if err := env.Duck.DB.QueryRowContext(ctx, fmt.Sprintf(
